+++ src/types/actions.ts
@@ -6,10 +6,12 @@
 import React from 'react';
 
 import { useModal, useSnackbar } from '@zextras/carbonio-design-system';
+import { TFunction } from 'i18next';
 
 import { EventType } from './event';
 import { Invite } from './store/invite';
-import { Folders } from '../carbonio-ui-commons/types/folder';
+import { useHistoryNavigation } from '../carbonio-ui-commons/helpers/use-history-navigation';
+import { Folders } from '../carbonio-ui-commons/types';
 import { Tags } from '../carbonio-ui-commons/types/tags';
 import { EventActionsId } from '../constants/event-actions';
 import { AppDispatch } from '../store/redux';
@@ -26,7 +28,8 @@ export type ActionsContext = {
 	closeModal: ReturnType<typeof useModal>['closeModal'];
 	createSnackbar: ReturnType<typeof useSnackbar>;
 	dispatch: AppDispatch;
-	replaceHistory?: (arg: any) => void;
+	replaceHistory: ReturnType<typeof useHistoryNavigation>['replaceHistory'];
+	t: TFunction;
 	tags: Tags;
 	onClose?: () => void;
 	panel?: boolean;