+++ src/types/store/invite.ts
@@ -34,7 +34,7 @@ export type Attendee = {
 export type InviteClass = 'PUB' | 'PRI' | 'CON';
 /* flags = 'u' | 'f' | 'a' | 'r' | 's' | 'w' | 'v' | 'd' | 'x' | 'n' | '!' | '?' | '+' */
 
-export type InviteFreeBusy = 'F' | 'B' | 'T' | 'U';
+export type InviteFreeBusy = 'F' | 'B' | 'T' | 'O';
 export type InviteOrganizer = {
 	a: string;
 	d: string;
@@ -163,7 +163,7 @@ export type Invite = {
 	location: string;
 	name: string;
 	noBlob: boolean;
-	organizer: InviteOrganizer;
+	organizer: InviteOrganizer; // FIXME: this actually can be undefined (see CO-1885)
 	recurrenceRule: any;
 	isRespRequested: boolean;
 	start: InviteDateFormat;