+++ src/view/event-panel-view/reply-buttons-part.tsx
@@ -14,7 +14,9 @@ import {
 	declineInvitation,
 	proposeNewTimeFn
 } from '../../actions/appointment-actions-fn';
+import { useHistoryNavigation } from '../../carbonio-ui-commons/helpers/use-history-navigation';
 import { useFoldersMap } from '../../carbonio-ui-commons/store/zustand/folder';
+import { PARTICIPATION_STATUS } from '../../constants/api';
 import { useAppDispatch } from '../../store/redux/hooks';
 import { EventType } from '../../types/event';
 import { Invite } from '../../types/store/invite';
@@ -26,15 +28,18 @@ type ReplyButtonProps = {
 
 export const ReplyButtonsPart = ({ event, invite }: ReplyButtonProps): ReactElement => {
 	const [t] = useTranslation();
+	const { replaceHistory } = useHistoryNavigation();
 	const dispatch = useAppDispatch();
 	const folders = useFoldersMap();
 	const context = useMemo(
 		() => ({
 			dispatch,
 			folders,
+			t,
+			replaceHistory,
 			isInstance: !!event.resource.ridZ
 		}),
-		[dispatch, event.resource.ridZ, folders]
+		[dispatch, event.resource.ridZ, folders, replaceHistory, t]
 	);
 
 	return (
@@ -53,7 +58,7 @@ export const ReplyButtonsPart = ({ event, invite }: ReplyButtonProps): ReactElem
 				icon="CheckmarkOutline"
 				color="success"
 				onClick={acceptInvitation({ event, invite, context })}
-				disabled={event.resource.participationStatus === 'AC'}
+				disabled={event.resource.participationStatus === PARTICIPATION_STATUS.ACCEPTED}
 			/>
 			<Padding horizontal="small" />
 			<Button
@@ -62,7 +67,7 @@ export const ReplyButtonsPart = ({ event, invite }: ReplyButtonProps): ReactElem
 				icon="QuestionMarkOutline"
 				color="warning"
 				onClick={acceptAsTentative({ event, invite, context })}
-				disabled={event.resource.participationStatus === 'TE'}
+				disabled={event.resource.participationStatus === PARTICIPATION_STATUS.TENTATIVE}
 			/>
 			<Padding horizontal="small" />
 			<Button
@@ -71,7 +76,7 @@ export const ReplyButtonsPart = ({ event, invite }: ReplyButtonProps): ReactElem
 				icon="CloseOutline"
 				color="error"
 				onClick={declineInvitation({ event, invite, context })}
-				disabled={event.resource.participationStatus === 'DE'}
+				disabled={event.resource.participationStatus === PARTICIPATION_STATUS.DECLINED}
 			/>
 			<Padding horizontal="small" />
 			<Button